refactor and chat completions
