+++ docs/plan/refactor_mcp_server_request.md
@@ -1,0 +1,3 @@
+Verify this plan, prepare refactor_mcp_server_execution.md file with checkboxes for your tasks, an execute on server.py, creating a new server with split components design supporting chat-completions and responses api.
+When you finish a task, mark the checkbox as checked.
+If you find new tasks - add them under a new header for later.
